+++ affinda/models/_models_py3.py
@@ -3512,6 +3512,8 @@ class DocumentUpdate(msrest.serialization.Model):
     :vartype is_archived: bool
     :ivar language: Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese.
     :vartype language: str
+    :ivar identifier: Specify a custom identifier for the document.
+    :vartype identifier: str
     """
 
     _attribute_map = {
@@ -3522,6 +3524,7 @@ class DocumentUpdate(msrest.serialization.Model):
         "is_rejected": {"key": "isRejected", "type": "bool"},
         "is_archived": {"key": "isArchived", "type": "bool"},
         "language": {"key": "language", "type": "str"},
+        "identifier": {"key": "identifier", "type": "str"},
     }
 
     def __init__(
@@ -3534,6 +3537,7 @@ class DocumentUpdate(msrest.serialization.Model):
         is_rejected: Optional[bool] = None,
         is_archived: Optional[bool] = None,
         language: Optional[str] = None,
+        identifier: Optional[str] = None,
         **kwargs,
     ):
         """
@@ -3552,6 +3556,8 @@ class DocumentUpdate(msrest.serialization.Model):
         :paramtype is_archived: bool
         :keyword language: Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese.
         :paramtype language: str
+        :keyword identifier: Specify a custom identifier for the document.
+        :paramtype identifier: str
         """
         super(DocumentUpdate, self).__init__(**kwargs)
         self.collection = collection
@@ -3561,6 +3567,7 @@ class DocumentUpdate(msrest.serialization.Model):
         self.is_rejected = is_rejected
         self.is_archived = is_archived
         self.language = language
+        self.identifier = identifier
 
 
 class DocumentWarning(msrest.serialization.Model):